+++ vite.config.ts
@@ -3,7 +3,7 @@ import preact from '@preact/preset-vite'
 import svgr from '@svgr/rollup'
 import path from 'path'
 import { crx } from '@crxjs/vite-plugin'
-import manifest from './manifest.json'
+import manifest from './manifest'
 
 // https://vitejs.dev/config/
 export default defineConfig({